requirements.txt
# Rise Local Lead Creation Pipeline - Dependencies
# Install: pip install -r requirements.txt

# HTTP Client
httpx>=0.25.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Supabase Client
supabase>=2.0.0

# Anthropic (Claude API)
anthropic>=0.18.0

# Google API Clients (for Sheets integration)
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# CSV Processing
pandas>=2.0.0

# Async Support
aiofiles>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
